Revert "Optimize memory footprint of line buffers"
This reverts commit 7ab6c92e18d468968811256e808b02309c160a22.
We need 32 bits for real color support.

+++ st.c
@@ -191,8 +191,8 @@ typedef XftColor Color;
 typedef struct {
 	Rune u;           /* character code */
 	ushort mode;      /* attribute flags */
-	ushort fg;        /* foreground  */
-	ushort bg;        /* background  */
+	uint32_t fg;      /* foreground  */
+	uint32_t bg;      /* background  */
 } Glyph;
 
 typedef Glyph *Line;
